requirements.txt
Bio==1.6.2
biopython==1.81
PyYAML==6.0.1
pandas==1.4.0
seaborn==0.11.2
matplotlib==3.5.1
numpy==1.22.0
